* compiling fixes
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@2806 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ luascript.cpp
@@ -1068,7 +1068,7 @@ int64_t LuaScriptInterface::popNumber(lua_State* L)
 	if(lua_isboolean(L, 0))
 		return lua_toboolean(L, 0);
 
-	return lua_tonumber(L, 0);
+	return (int64_t)lua_tonumber(L, 0);
 }
 
 double LuaScriptInterface::popFloatNumber(lua_State* L)
